open new instance to other package other improvements
